moved structured data stuff into separate file. at least one case with a pydantic model worked
