+++ archytas/constants.py
@@ -1,0 +1,5 @@
+"""
+Constants shared across the package.
+"""
+
+TAB = '    '
